+++ src/Pages/Sender/EnterTime.tsx
@@ -64,7 +64,6 @@ const EnterTime: React.FC = () => {
               justifyContent: 'center',
               alignItems: 'center',
               width: '80vw',
-              marginLeft: '10vw',
             }}
           />
         </Form.Item>
